requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
